Moved #922 ticket tests to renderer file.
